feat(pages): Challenge page shows different views depending on screen size
Custom hook 'useWindowSize' added which returns a object containing the window width and height. Thiis is updated when the window is resized. 'useWindowSize' is then used to determine which view to render to the user. Desktop users will have a side-by-side view of the challenge and code editor. Mobile users will only see the code editor and a menu button which opens a drawer containing the challenge data.